requirements.txt
boto3==1.22.1
botocore==1.25.1
DateTime==4.4
Flask==2.0.3
mysql-connector-python==8.0.28
python-dotenv==0.19.2
requests==2.22.0
requests-unixsocket==0.2.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3